agent/pyproject.toml
[project]
name = "agent-robert"
version = "0.1.0"
description = "Ultra-minimal, security-first AI assistant"
requires-python = ">=3.11"
license = {text = "MIT"}
authors = [
    {name = "Student team"}
]

dependencies = [
    "httpx>=0.25.0",
    "python-dotenv>=1.0.0",
    "typer>=0.9.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "ruff>=0.1.0",
]

[project.scripts]
robert = "robert.main:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["robert"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W"]
ignore = ["E501"]
